+++ SQLiteStudio3/guiSQLiteStudio/dialogs/configdialog.h
@@ -69,6 +69,8 @@ class GUI_API_EXPORT ConfigDialog : public QDialog
         void initLangs();
         void initTooltips();
         void initColors();
+        void updateColorsAfterLoad();
+        void toggleColorButtonState(CfgEntry* colorCheckEntry);
         void applyStyle(QWidget* widget, QStyle* style);
         QTreeWidgetItem* getPluginsCategoryItem() const;
         QTreeWidgetItem* getPluginsCategoryItem(PluginType* type) const;