+++ SQLiteStudio3/guiSQLiteStudio/style.h
@@ -9,6 +9,8 @@ class CfgEntry;
 
 class Style : public QProxyStyle
 {
+    Q_OBJECT
+
     public:
         static Style* getInstance();
         static bool isDark(const QStyle* style);
@@ -28,6 +30,9 @@ class Style : public QProxyStyle
 
         ExtendedPalette extPalette;
         QPalette initialPalette;
+
+    signals:
+        void paletteChanged();
 };
 
 #define STYLE Style::getInstance()